+++ dynamic_forms/models.py
@@ -53,6 +53,8 @@ class FormModel(models.Model):
         help_text=_('Allow a user to view the input at a later time. This '
             'requires the “Store in database” action to be active. The sender '
             'will be given a unique URL to recall the data.'))
+    recipient_email = models.EmailField(_('Recipient email'), blank=True,
+        null=True, help_text=_('Email address to send form data.'))
 
     class Meta:
         ordering = ['name']