+++ docs/source/changes.rst
@@ -9,6 +9,7 @@ v0.4 (under development)
 
 * Added support for Django 1.8 and experimental support for Django 1.9. (#19)
 * Removed ``django-appconf`` dependency.
+* Added per form email receivers (thanks to Nar Chhantyal). (#13)
 
 
 v0.3.4