+++ docs/source/changes.rst
@@ -4,8 +4,21 @@ Changes
 
 .. py:currentmodule:: dynamic_forms
 
-v0.4 (under development)
+v0.5 (under development)
 ========================
+
+.. warning::
+
+   **django-dynamic-forms** 0.5.x will only support Django >= 1.7!
+
+
+v0.4
+====
+
+.. warning::
+
+   **django-dynamic-forms** 0.4.x will be the latest version branch that
+   supports Django < 1.7!
 
 * Added support for Django 1.8 and experimental support for Django 1.9. (#19)
 * Removed ``django-appconf`` dependency.